# backend/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Personal Activity Manager API",
    "description": "Documentação da API"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/user/": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Endpoint that returns all users",
        "responses": {
          "200": {
            "description": "Users found",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint that creates a new user",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "User data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddUser"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created successfully",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user/{id}": {
      "put": {
        "tags": [
          "User"
        ],
        "description": "Endpoint that updates a user by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the user"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "Updated user data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User updated successfully",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Invalid ID",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Invalid user ID"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "User not found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "User not found"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "description": "Endpoint that deletes a user by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the user"
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Invalid ID",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Invalid user ID"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "User not found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "User not found"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "first_name": {
          "type": "string",
          "example": "Luiz Felippe"
        },
        "last_name": {
          "type": "string",
          "example": "Brizola"
        },
        "password": {
          "type": "string",
          "example": "!22lf"
        },
        "status": {
          "type": "string",
          "example": "true"
        }
      }
    },
    "AddUser": {
      "type": "object",
      "properties": {
        "first_name": {
          "type": "string",
          "example": "Luiz Felippe"
        },
        "last_name": {
          "type": "string",
          "example": "Brizola"
        },
        "password": {
          "type": "string",
          "example": "!22lf"
        }
      },
      "required": [
        "first_name",
        "last_name",
        "password"
      ]
    },
    "UpdateUser": {
      "type": "object",
      "properties": {
        "first_name": {
          "type": "string",
          "example": "Luiz Felippe"
        },
        "last_name": {
          "type": "string",
          "example": "Brizola"
        },
        "password": {
          "type": "string",
          "example": "!22lf"
        },
        "status": {
          "type": "string",
          "example": "true"
        }
      }
    }
  }
}